+++ docs/changes.md
@@ -9,6 +9,7 @@
 - Fixed broken Apple Lisa fonts.
 - Fixed the local file viewer toggling on non-text file tabs when the file's extension is in uppercase.
 - Fixed certain fonts that use spaces within their names, breaking the loading text files.
+- Fixed broken SAUCE IBM VGA50 font.
 - Improved support and display of multiline SAUCE comments.
 - Improve the positioning of the error alert box.
 - Spelling and typo fixes for both UK and US English.